Robot/src/test/java/au/com/amdocs/executer/factory/RightCommandExecuterFactoryTest.java
package au.com.amdocs.executer.factory;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

import au.com.amdocs.command.interf.CommandExecuter;
import au.com.amdocs.command.interf.ExecuterFactory;
import au.com.amdocs.executer.RightCommandExecuter;
import au.com.amdocs.executer.factory.RightCommandExecuterFactory;

public class RightCommandExecuterFactoryTest {

	@Test
	public void testBuildExecuter() {
		ExecuterFactory factory = new RightCommandExecuterFactory();
		CommandExecuter executer = factory.buildExecuter(null, null);
		assertNotNull(executer);
		assertTrue(executer instanceof RightCommandExecuter);
	}
}
